+++ Includes/ManualRevalidation.php
@@ -7,7 +7,7 @@
  *
  * @package RevalidatePosts
  * @since 1.4.0
- * @version 1.5.0
+ * @version 1.8.0
  * @author Silver Assist
  * @license Polyform Noncommercial 1.0.0
  */